moved kprobe attached method test
