+++ pkg/network/tracer/tracer_linux_test.go
@@ -28,6 +28,8 @@ import (
 
 	"golang.org/x/sys/unix"
 
+	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
+	"github.com/DataDog/datadog-agent/pkg/network/config"
 	"github.com/DataDog/datadog-agent/pkg/network/http"
 	nettestutil "github.com/DataDog/datadog-agent/pkg/network/testutil"
 	tracertest "github.com/DataDog/datadog-agent/pkg/network/tracer/testutil"
@@ -1650,4 +1652,25 @@ func TestShortWrite(t *testing.T) {
 	}, 3*time.Second, 500*time.Millisecond, "couldn't find connection used by short write")
 
 	assert.Equal(t, sent, conn.MonotonicSum().SentBytes)
+}
+
+func TestKprobeAttachWithKprobeEvents(t *testing.T) {
+	cfg := config.New()
+	cfg.AttachKprobesWithKprobeEventsABI = true
+
+	tr, err := NewTracer(cfg)
+	require.NoError(t, err)
+	defer tr.Stop()
+
+	cmd := []string{"curl", "-k", "-o/dev/null", "facebook.com"}
+	exec.Command(cmd[0], cmd[1:]...).Run()
+
+	stats := ddebpf.GetProbeStats()
+	require.NotNil(t, stats)
+
+	p_tcp_sendmsg, ok := stats["p_tcp_sendmsg_hits"]
+	require.True(t, ok)
+	fmt.Printf("p_tcp_sendmsg_hits = %d\n", p_tcp_sendmsg)
+
+	assert.Greater(t, p_tcp_sendmsg, int64(0))
 }
