+++ pkg/network/tracer/tracer_test.go
@@ -36,7 +36,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"golang.org/x/sync/errgroup"
 
-	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/network"
 	"github.com/DataDog/datadog-agent/pkg/network/config"
 	"github.com/DataDog/datadog-agent/pkg/network/http"
@@ -1844,27 +1843,6 @@ func TestRuntimeCompilerEnvironmentVar(t *testing.T) {
 	enabled := os.Getenv(runtimeCompilationEnvVar) != ""
 	assert.Equal(t, enabled, cfg.EnableRuntimeCompiler)
 	assert.NotEqual(t, enabled, cfg.AllowPrecompiledFallback)
-}
-
-func TestKprobeAttachWithKprobeEvents(t *testing.T) {
-	cfg := config.New()
-	cfg.AttachKprobesWithKprobeEventsABI = true
-
-	tr, err := NewTracer(cfg)
-	require.NoError(t, err)
-	defer tr.Stop()
-
-	cmd := []string{"curl", "-k", "-o/dev/null", "facebook.com"}
-	exec.Command(cmd[0], cmd[1:]...).Run()
-
-	stats := ddebpf.GetProbeStats()
-	require.NotNil(t, stats)
-
-	p_tcp_sendmsg, ok := stats["p_tcp_sendmsg_hits"]
-	require.True(t, ok)
-	fmt.Printf("p_tcp_sendmsg_hits = %d\n", p_tcp_sendmsg)
-
-	assert.Greater(t, p_tcp_sendmsg, int64(0))
 }
 
 func testConfig() *config.Config {
